User commit

+++ src/main/java/com/example/demo/repository/StudentRepository.java
@@ -2,6 +2,6 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpRepository;
 import com.example.demo.entity.Student;
 @Repository
-public interface StudentRepository extends JpRepository<Student,Long>{
+public interface StudentRepository extends JpRepository<Student,Integer>{
 
 }
